Final - with corrections made
A final version of the website was submitted on February 12th. There
were a few  unresolved issues; I went in and made slight modifications.
- I do not expected to be graded on the the Feb 13 changes:
Resolved the navigation issue and added a frame with drop shadow around
the body and moved the <em> to the right on the success stories page.